reworked  Wall
