+++ src/compas_timber/connections/plate_joint.py
@@ -14,71 +14,6 @@ from .joint import JointTopology
 from .solver import PlateConnectionSolver
 
 
-class InterfaceRole(object):
-    """
-    Enumeration of the possible interface roles.
-
-    Attributes
-    ----------
-    MAIN : literal("MAIN")
-        The interface is the main interface.
-    CROSS : literal("CROSS")
-        The interface is the cross interface.
-    NONE : literal("NONE")
-        The interface has no specific role. E.g. when a miter joint is used.
-    """
-
-    MAIN = "MAIN"
-    CROSS = "CROSS"
-    NONE = "NONE"
-
-
-class PlateToPlateInterface(object):
-    """
-    polyline : :class:`compas.geometry.Polyline`
-        The outline of the interface area.
-    frame : :class:`compas.geometry.Frame`
-        The frame of the interface area.
-        xaxis : interface normal (towards other plate)
-        yaxis : up along the interface side
-        normal: width direction, perpendicular to the interface
-    topology : literal(JointTopology)
-        The topology of the joint in which the interface is used.
-        E.g. L or T
-    interface_role : literal(InterfaceRole)
-        The role of the interface in the joint.
-
-    """
-
-    def __init__(self, polyline, frame, edge_index, topology, interface_role=None):
-        self.polyline = polyline
-        self.frame = frame
-        self.edge_index = edge_index  # index of the edge in the plate outline where the interface is located
-        self.topology = topology  # TODO: don't like this here
-        self.interface_role = interface_role if interface_role else InterfaceRole.NONE
-
-    def __repr__(self):
-        return "PlateToPlateInterface({0}, {1})".format(
-            self.interface_role,
-            JointTopology.get_name(self.topology),
-        )
-
-    def as_plane(self):
-        """Returns the interface as a plane.
-
-        Returns
-        -------
-        :class:`compas.geometry.Plane`
-            The plane of the interface.
-        """
-        return Plane.from_frame(self.frame)
-
-    @property
-    def width(self):
-        """Returns the width of the interface polyline."""
-        return distance_line_line(self.polyline.lines[0], self.polyline.lines[2])
-
-
 class PlateJoint(Joint):
     """Models a plate to plate interaction.
 
@@ -105,10 +40,7 @@ class PlateJoint(Joint):
         The second plate.
     plates : tuple of :class:`compas_timber.elements.Plate`
         The plates that are connected.
-    interface_a : :class:`compas.geometry.PlanarSurface`
-        The interface surface of plate_a where it meets plate_b.
-    interface_b : :class:`compas.geometry.PlanarSurface`
-        The interface surface of plate_b where it meets plate_a.
+
 
     """
 
@@ -150,57 +82,7 @@ class PlateJoint(Joint):
     def elements(self):
         return self.plate_a, self.plate_b
 
-    @property
-    def geometry(self):
-        return self.interface_a.polyline
-
-    @property
-    def interfaces(self):
-        return self.interface_a, self.interface_b
-
-    @property
-    def interface_a(self):
-        a_interface_polyline = Polyline(
-            [
-                self.a_outlines[0][self.a_segment_index],
-                self.a_outlines[0][self.a_segment_index + 1],
-                self.a_outlines[1][self.a_segment_index + 1],
-                self.a_outlines[1][self.a_segment_index],
-                self.a_outlines[0][self.a_segment_index],
-            ]
-        )
-
-        frame = Frame.from_points(a_interface_polyline.points[0], a_interface_polyline.points[1], a_interface_polyline.points[-2])
-        if dot_vectors(frame.normal, Vector.from_start_end(self.b_planes[1].point, self.b_planes[0].point)) < 0:
-            frame = Frame.from_points(a_interface_polyline.points[1], a_interface_polyline.points[0], a_interface_polyline.points[2])
-        return PlateToPlateInterface(
-            a_interface_polyline,
-            frame,
-            self.a_segment_index,
-            self.topology,
-        )
-
-    @property
-    def interface_b(self):
-        b_interface_polyline = Polyline(
-            [
-                self.a_outlines[1][self.a_segment_index],
-                self.a_outlines[1][self.a_segment_index + 1],
-                self.a_outlines[0][self.a_segment_index + 1],
-                self.a_outlines[0][self.a_segment_index],
-                self.a_outlines[1][self.a_segment_index],
-            ]
-        )
-        frame = Frame.from_points(b_interface_polyline.points[0], b_interface_polyline.points[1], b_interface_polyline.points[-2])
-        if dot_vectors(frame.normal, Vector.from_start_end(self.b_planes[0].point, self.b_planes[1].point)) < 0:
-            frame = Frame.from_points(b_interface_polyline.points[1], b_interface_polyline.points[0], b_interface_polyline.points[2])
-        return PlateToPlateInterface(
-            b_interface_polyline,
-            frame,
-            self.b_segment_index,
-            self.topology,
-        )
-
+   
     def calculate_topology(self, allow_reordering=False):
         """Calculate the topology of the joint based on the plates."""
         topo_results = PlateConnectionSolver().find_topology(self.plate_a, self.plate_b)
@@ -241,23 +123,17 @@ class PlateJoint(Joint):
         kwargs.update({"a_segment_index": candidate.a_segment_index, "b_segment_index": candidate.b_segment_index})  # pass segment indices from candidate
         return super(PlateJoint, cls).promote_joint_candidate(model, candidate, reordered_elements=reordered_elements, **kwargs)
 
-    def add_features(self):
-        """Add features to the plates based on the joint."""
+    def add_extensions(self):
+        """Adjusts plate outlines to outer shape required for the joint."""
         if self.plate_a and self.plate_b:
             if self.topology is None or (self.a_segment_index is None and self.b_segment_index is None):
                 self.calculate_topology()
             self.reorder_planes_and_outlines()
-            self._adjust_plate_outlines()
-            self.plate_a.add_interface(self.interface_a)
-            self.plate_b.add_interface(self.interface_b)
+            self._adjust_plate_outlines() 
 
-    def get_interface_for_plate(self, plate):
-        if plate is self.plate_a:
-            return self.interface_a
-        elif plate is self.plate_b:
-            return self.interface_b
-        else:
-            raise ValueError("Plate not part of this joint.")
+    def add_features(self):
+        """Adds features to the plates based on the joint. this should be implemented in subclasses if needed."""
+        pass
 
     def reorder_planes_and_outlines(self):
         if dot_vectors(self.plate_b.frame.normal, get_polyline_segment_perpendicular_vector(self.plate_a.outline_a, self.a_segment_index)) < 0:
@@ -273,14 +149,6 @@ class PlateJoint(Joint):
             if dot_vectors(self.plate_a.frame.normal, get_polyline_segment_perpendicular_vector(self.plate_b.outline_a, self.b_segment_index)) < 0:
                 self.a_planes = self.plate_a.planes[::-1]
                 self.a_outlines = self.plate_a.outlines[::-1]
-
-    def restore_beams_from_keys(self, *args, **kwargs):
-        # TODO: this is just to keep the peace. change once we know where this is going.
-        self.restore_plates_from_keys(*args, **kwargs)
-
-    def restore_plates_from_keys(self, model):
-        self.plate_a = model.element_by_guid(self.plate_a_guid)
-        self.plate_b = model.element_by_guid(self.plate_b_guid)
 
     def flip_roles(self):
         self.plate_a, self.plate_b = self.plate_b, self.plate_a
