+++ src/compas_timber/elements/__init__.py
@@ -1,7 +1,7 @@
 from .beam import Beam
 from .plate import Plate
-from .slab import Opening
-from .slab import OpeningType
+from .wall import Opening
+from .wall import OpeningType
 from .slab import Slab
 from .wall import Wall
 from .fastener import Fastener
