+++ tests/compas_timber/test_plate_joint.py
@@ -17,6 +17,7 @@ def test_simple_joint_and_reset():
 
     kwargs = {"topology": JointTopology.TOPO_EDGE_EDGE, "a_segment_index": 1, "b_segment_index": 0}
     joint = PlateMiterJoint(plate_a, plate_b, **kwargs)
+    joint.add_extensions()
     joint.add_features()
     assert isinstance(joint, PlateMiterJoint), "Expected joint to be a PlateMiterJoint"
     assert any([plate_a.outline_a.points[i] != polyline_a.points[i] for i in range(len(plate_a.outline_a.points))]), "Expected joint to change outline_a"
@@ -32,6 +33,7 @@ def test_simple_joint_and_reset_no_kwargs():
     plate_b = Plate.from_outline_thickness(Polyline(polyline_b.points), 1)
 
     joint = PlateMiterJoint(plate_a, plate_b)
+    joint.add_extensions()
     joint.add_features()
     assert isinstance(joint, PlateMiterJoint), "Expected joint to be a PlateMiterJoint"
     assert any([plate_a.outline_a.points[i] != polyline_a.points[i] for i in range(len(plate_a.outline_a.points))]), "Expected joint to change outline_a"
